SectionManager: Make interactor & log_printer global
Earlier logging objects namely outputter and log_printer were set
individually in each section. The retrieve_logging_objects() method
has been moved from Section class to SectionManager class and thus
outputter and log_printer are set globally in SectionManager. They take
their respective settings from the default section of the coala run.
This commit is part of a series of commits which implements the
feature.

See https://github.com/coala-analyzer/coala/issues/386
