+++ coalib/settings/Section.py
@@ -1,14 +1,9 @@
 from collections import OrderedDict
 import copy
-from coalib.output.NullInteractor import NullInteractor
 
-from coalib.output.printers.FilePrinter import FilePrinter
-from coalib.output.printers.LOG_LEVEL import LOG_LEVEL
 from coalib.output.printers.LogPrinter import LogPrinter
-from coalib.output.printers.NullPrinter import NullPrinter
 from coalib.output.ConsoleInteractor import ConsoleInteractor, ConsolePrinter
 from coalib.output.Interactor import Interactor
-from coalib.misc.i18n import _
 from coalib.settings.Setting import Setting
 
 
@@ -55,42 +50,6 @@ class Section:
             return bool(self.get("enabled", "true"))
 
         return self.name.lower() in targets
-
-    def retrieve_logging_objects(self):
-        """
-        Creates an appropriate log printer and interactor according to the
-        settings.
-        """
-        log_type = str(self.get("log_type", "console")).lower()
-        output_type = str(self.get("output", "console")).lower()
-        str_log_level = str(self.get("log_level", "")).upper()
-        log_level = LOG_LEVEL.str_dict.get(str_log_level, LOG_LEVEL.WARNING)
-
-        if log_type == "console":
-            self.log_printer = ConsolePrinter(log_level=log_level)
-        else:
-            try:
-                # ConsolePrinter is the only printer which may not throw an
-                # exception (if we have no bugs though) so well fallback to him
-                # if some other printer fails
-                if log_type == "none":
-                    self.log_printer = NullPrinter()
-                else:
-                    self.log_printer = FilePrinter(filename=log_type,
-                                                   log_level=log_level)
-            except:
-                self.log_printer = ConsolePrinter(log_level=log_level)
-                self.log_printer.log(
-                    LOG_LEVEL.WARNING,
-                    _("Failed to instantiate the logging method '{}'. Falling "
-                      "back to console output.").format(log_type))
-
-        if output_type == "none":
-            self.interactor = NullInteractor(log_printer=self.log_printer)
-        else:
-            self.interactor = ConsoleInteractor.from_section(
-                self,
-                log_printer=self.log_printer)
 
     def append(self, setting, custom_key=None):
         if not isinstance(setting, Setting):
